pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "board-md"
version = "0.1.0"
description = "Markdown-native project board for the AI agent era"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
dependencies = [
    "click>=8.0",
    "python-frontmatter>=1.1",
]

[project.scripts]
board = "board_md.cli:cli"

[project.urls]
Homepage = "https://github.com/gongwu-ai/board.md"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
